requirements.txt
confz==1.8.1
cryptography==40.0.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
numpy==1.24.2
pandas==1.5.3
pydantic==1.10.7
pyjwt==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlalchemy==1.4.41
sqlmodel==0.0.8
typing_extensions==4.5.0
typer==0.7.0
uvicorn==0.21.1
